+++ Readme_Changes.txt
@@ -1,0 +1,11 @@
+Changes Made:
+
+Sem hash functions are purposely kept inside the .py file and new class has not been created to match the file name ohterwise such addittional change can be made to put them in once class in anoter file
+
+Class MeraDatset is transferred to a new file to reduce clutter
+Functions clubbed together to improve clarity and testing ability of individual function if assigned individual cell in Jupyter Notebook or any other .ipynb
+Reduced the number of lines left
+Comments also mentioned above the changes Made
+read_csv is transferred to new file to increase reusability
+plot_results() is transferred to new file to increase reusability
+tokenize, preprocess, data_for_training, get_vectorizer moved to creating_data_train_test.py to improve readibilty and reusability
