+++ plot_result.py
@@ -1,0 +1,27 @@
+import matplotlib.pyplot as plt
+def plot_results(results):
+    # make some plots
+    indices = np.arange(len(results))
+
+    results = [[x[i] for x in results] for i in range(4)]
+
+    clf_names, score, training_time, test_time = results
+    training_time = np.array(training_time) / np.max(training_time)
+    test_time = np.array(test_time) / np.max(test_time)
+
+    plt.figure(figsize=(12, 8))
+    plt.title("Score")
+    plt.barh(indices, score, .2, label="score", color='navy')
+    plt.barh(indices + .3, training_time, .2, label="training time",
+             color='c')
+    plt.barh(indices + .6, test_time, .2, label="test time", color='darkorange')
+    plt.yticks(())
+    plt.legend(loc='best')
+    plt.subplots_adjust(left=.25)
+    plt.subplots_adjust(top=.95)
+    plt.subplots_adjust(bottom=.05)
+
+    for i, c in zip(indices, clf_names):
+        plt.text(-.3, i, c)
+
+    plt.show()
